refactor: move home page into folder and break out components into separate files
